+++ src/main/java/org/owasp/webgoat/lessons/insecurelogin/InsecureLoginTask.java
@@ -20,7 +20,7 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.insecure_login;
+package org.owasp.webgoat.lessons.insecurelogin;
 
 import org.owasp.webgoat.container.assignments.AssignmentEndpoint;
 import org.owasp.webgoat.container.assignments.AttackResult;
