+++ src/main/resources/lessons/insecurelogin/html/InsecureLogin.html
@@ -6,12 +6,12 @@
         <!-- reuse this lesson-page-wrapper block for each 'page' of content in your lesson -->
         <!-- include content here. Content will be presented via asciidocs files,
         which you put in src/main/resources/plugin/lessonplans/{lang}/{fileName}.adoc -->
-        <div class="adoc-content" th:replace="doc:lessons/insecure_login/documentation/InsecureLogin_Intro.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/insecurelogin/documentation/InsecureLogin_Intro.adoc"></div>
     </div>
 
     <div class="lesson-page-wrapper">
         <!-- stripped down without extra comments -->
-        <div class="adoc-content" th:replace="doc:lessons/insecure_login/documentation/InsecureLogin_Task.adoc"></div>
+        <div class="adoc-content" th:replace="doc:lessons/insecurelogin/documentation/InsecureLogin_Task.adoc"></div>
         <div class="attack-container">
             <div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
             <script th:src="@{/lesson_js/credentials.js}"></script>
