+++ src/main/resources/lessons/insecurelogin/js/credentials.js
